test: add test case for validator

+++ src/main/java/org/cardanofoundation/explorer/consumercommon/validation/TxIndexValidator.java
@@ -8,6 +8,6 @@ public class TxIndexValidator implements ConstraintValidator<TxIndex, Short> {
 
   @Override
   public boolean isValid(Short aShort, ConstraintValidatorContext constraintValidatorContext) {
-    return aShort >= BigInteger.ZERO.intValue();
+    return aShort >= BigInteger.ZERO.shortValue();
   }
 }
