+++ src/main/java/org/cardanofoundation/explorer/consumercommon/validation/Word63TypeValidator.java
@@ -15,6 +15,6 @@ public class Word63TypeValidator implements ConstraintValidator<Word63Type, Long
    */
   @Override
   public boolean isValid(Long aLong, ConstraintValidatorContext constraintValidatorContext) {
-    return aLong >= BigInteger.ZERO.intValue();
+    return aLong >= BigInteger.ZERO.longValue();
   }
 }
